remove pages and rename services
